--- app/botevents.ts
@@ -1,3 +1,0 @@
-import EventEmitter from "events";
-const bot_events = new EventEmitter();
-export default bot_events;
