+++ app/eventemitters/bot-worker.eventemitter.ts
@@ -1,0 +1,3 @@
+import EventEmitter from "events";
+const bot_worker_ee = new EventEmitter();
+export default bot_worker_ee;
