+++ app/eventemitters/manager-controller.eventemitter.ts
@@ -1,0 +1,3 @@
+import EventEmitter from "events";
+const manager_controller_ee = new EventEmitter();
+export default manager_controller_ee;
