+++ app/interfaces/ai-client.interface.ts
