+++ app/interfaces/log-processing.interface.ts
