+++ app/interfaces/message.interface.ts
