+++ app/models/player.model.ts
@@ -1,4 +1,4 @@
-import { PlayerStats } from "./playerstats.ts";
+import { PlayerStats } from "./playerstats.model.ts";
 
 export class Player {
     private player_id: string;
