+++ app/models/playeraction.model.ts
@@ -1,4 +1,4 @@
-import { Action } from "../utils/logprocessing-utils.ts";
+import { Action } from "../utils/log-processing.util.ts";
 
 export class PlayerAction {
     private player_id: string;
