+++ app/utils/data-structures.util.ts
